kafka-handler/requirements.txt
fastapi==0.61.2
uvicorn==0.12.2
kafka-python
